Zoom camera in /sp, ignoring the ImGui config.
Now it is possible to zoom the camera in the spectator with the 'cfg.zoomCameraInSpectator' setting enabled. Management is carried out through the mouse wheel:
up - zooms out the camera
down - zooms in on the camera
Zoom in/zoom out through FOV within a range of 50 to 150. Doesn't conflict with 'cfg.additions.fisheye.use' in any way.

Now ImGui won't get windows settings.

Signed-off-by: Dmitry Lyubimcev <dzone@danwin1210.de>