MovableWindows.lua module, player checker and fps counter removed.
MovableWindows.lua: new module.
Allows the user to conveniently adjust the position of all windows that are defined in MovableWindows.init { ... }. When hovering over a window, displays a brief description of what, when pressed, changes the position of the window relative to `MouseDelta`. Pulls the window to the sides and corners of the screen if the window position is out of resolution. Magnetizes the window to the center (x and y) within a radius of 5 pixels.
For more details, see the documentation in the code.

gadmin.lua: Removed the old system for moving window positions, now the MovableWindows module has replaced it.

Signed-off-by: Dmitry Lyubimcev <dzone@danwin1210.de>