Fixes, implemented almost all features in `config.root.additions`
Implemented almost all features in `config.root.additions`:
    # additions.fishEye
    # additions.mention
    # additions.zoomSpectatorCamera
    # additions.tracers
    # additions.swapLayout
    # additions.idInKillList
    # additions.deathNotifyInChat
    # additions.autoLogin
At the moment, all that remains from this section is to implement `additions.autoAdminLogin`.

In addition to the changes above, an AirBreak has been implemented.

Fixes:
# Fixed `rgb2color(rgb)` local function in sampev.onServerMessage (located in src/events/onServerMessage.clue`.
# Fixed wrong flag value for control key in `onVehicleSync` and `onPassengerSync` synchronizations.

At the moment, in order to successfully close a pull request, all that remains is to implement requests to accept commands, automatic `/alogin` after authorization, and the AutoUpdate.clue module. Also update the documentation in the repository.

Signed-off-by: Dmitry Lyubimcev <dzone@danwin1210.de>