Fixes, additions.
Far chat:
- Now only those actions (by `/me` or `/ame`) sent by the
player are displayed.
- If the received message color is black, it will change to white.

Now, if a technical administrator sends a message to the
administrators’ chat, his nickname will be saved by the script.

Mentioning everyone now also reacts to `@all` in administrators'
chat.

Fixed that the online counter counted when the game was paused.

The function of drawing text in different colors in the one
string has been rewritten:
    method text::addColored =>
    method text::renderTextObjects(drawList, position, objects)

Fixed that when clicking on a nickname in the administrator's
list window, it did not send a notification about
the successful copying of the text.

The font in the display of information about the vehicle,
information about the player's weapon,
the WallHack and the administrator’s nickname
on the skin has been changed. (to Cantarell)

The format for displaying information about the car
has been changed:
    HP: <number> · <Opened|Locked> · <On|Off>
    <vehicle name>[<vehicle id>]

Now you can change the font size for
information about the vehicle,
information about the player's weapon,
the WallHacks and administrator nicknames on
the skin separately.

Now you can change the color of the nickname
that is drawn on the administrator’s skin (default = "FF8585").

Now the color of the icon in the notification will be
the same as the text (= `imgui.GetStyle().Colors[imgui.Col.Text]`).

--------------------------------------------
|`40be29c` commit (wip), parent = `e0fde9b`|
--------------------------------------------

Fixed that the color of the report did not change if
it was wrapped to 3 or more messages.

Now you can use binder variables in notes.

Fixed `${{ CLOSEST_PLAYER_ID }}` and `${{ CLOSEST_PLAYER_NICKNAME }}`
were cause script crash when user try to use some of them.

The triangle in the tooltip is now more visible.

Triangle smoothing is enabled in the tooltip.

Fixed that turning off wallhack with a hotkey did not work on the skin bones.

Added pop-up windows when you hover over online statistics and
answers in the main window: they show detailed statistics by day of the week.

Signed-off-by: Dmitry Lyubimcev <dzone@danwin1210.de>