fix(server messages, player's platform detecting): security
Fixed possible deletion of a message from a chat when its
content is changed by a script and corresponding buffer overflow.

    It is worth noting that the buffer sometimes overflows even if the script
    does not intervene, or if the script respects the maximum number of characters
    before the server carries the string.

Added a check for both types of player platform state being spectated by the user:
from computer or phone. If the information received from the server is invalid, the
player can be treated as connected from the phone.

Signed-off-by: Dmitry Lyubimcev <dzone@danwin1210.de>