feat: auto update system & gadmin-update-handler.asi
The auto-update system has been implemented. It consists of a new
shared library (update-handler) and an in-game display of release
details: name, version, download count, file size, and description,
which is rendered via ImGuiMarkdown library.

gadmin-update-handler.asi is a file for handling updates. When loaded,
it iterates over all files in the directory where it is loaded and
checks the resource file (if there is one) of each file for OriginalFilename.
When found, the loader sends a request to the GitHub API to get information about
the latest version. If the received version differs from the one in the resource
file (FileVersion), it will ask the user whether the update should be installed
or not. After the update, the new version of gadmin.asi will be loaded. The
implementation is done using MessageBoxW and on the first thread.

There is also a new static library that provides access to the internet_t class: common.

Signed-off-by: Dmitry Lyubimcev <dzone@danwin1210.de>